fix update rotate settings

+++ app/src/main/java/utb/dip/jp/simple24hclock/AppWidgetProps.kt
@@ -14,7 +14,7 @@ data class AppWidgetProps(
     val timezone: String? = "",
     var lat: Float = 0F,
     var lng: Float = 0F,
-    var rotate: Int = 0,
+    var rotate: Float = 0F,
 )
 
 internal fun getAppWidgetProps(prefs: SharedPreferences, id: Int): AppWidgetProps {
@@ -30,7 +30,7 @@ internal fun getAppWidgetProps(prefs: SharedPreferences, id: Int): AppWidgetProp
         timezone = prefs.getString("timezone_$id", ""),
         lat = prefs.getFloat("lat_$id", 0F),
         lng = prefs.getFloat("lng_$id", 0F),
-        rotate = prefs.getInt("rotate_$id", 0),
+        rotate = prefs.getFloat("rotate_$id", 0F),
     )
 }
 
@@ -47,6 +47,6 @@ internal fun putAppWidgetProps(editor: SharedPreferences.Editor, props: AppWidge
     editor.putString("timezone_$id", props.timezone)
     editor.putFloat("lat_$id", props.lat)
     editor.putFloat("lng_$id", props.lng)
-    editor.putInt("rotate_$id", props.rotate)
+    editor.putFloat("rotate_$id", props.rotate)
 
 }
